started writing and improved code base
TEA should run many times faster as a result.

Some numbers may have changed? I can’t figure out if that’s because of
the dictionary I’m using or something else, but I’m worried that there
was a bug prior. This new code is considerably cleaner and it looks
fine to me….